feat: ConfirmLocationViewController add
Added view that allows users to confirm their location on a map before it is posted to the API.The view is centered and zoomed-in on the location being added by the user. A button on a map is used to accomplish the confirmation. Text on the button changes to indicate post is made. View segues to MapViewController centering around the new location with a greater radius - or zoomed out.