fix/feat: replaced confirm location button with a floating version and connected action
Button did not show. The button was being added as a subview before the properties were configuring. Replaced the button with a a better formatted button that is animated calling attention to user.

Separated add location action into two steps - find the location which shows the geolocated location on a separate view and confirm location which allows user to add location from that view.